When device width is less than 760, loads mobile radio component
