+++ src/components/RadioPage/Mobile/MobileRadio.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+class MobileRadio extends React.Component {
+    constructor(props) {
+        super(props);
+    }
+
+    render() {
+        return (
+            <h1>Yooooooooo</h1>
+        )
+    }
+}
+
+export default MobileRadio
